pyproject.toml
[project]
name = "revenueforecast"
version = "0.1.0"
description = "Local OCR pipeline: docs -> SQLite -> Excel/dashboard via llama.cpp"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.7",
    "pillow>=10.3",
    "pypdfium2>=4.30",
    "pandas>=2.2",
    "openpyxl>=3.1",
    "streamlit>=1.36",
    "plotly>=5.22",
    "tomli>=2.0; python_version < '3.11'",
    "pywebview>=5.0",
    "pyobjc-core>=10.0; sys_platform == 'darwin'",
    "pyobjc-framework-cocoa>=10.0; sys_platform == 'darwin'",
    "pyobjc-framework-webkit>=10.0; sys_platform == 'darwin'",
]

[project.scripts]
rf-ingest = "src.pipeline:cli_ingest"
rf-export = "src.export:cli_export"
rf-dashboard = "src.dashboard:cli_dashboard"
rf-desktop = "src.desktop:cli_desktop"

[dependency-groups]
dev = ["pytest>=8.2"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
